+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/keystonebirthday/height/KeystoneBDHeightArgs.kt
@@ -1,0 +1,9 @@
+package co.electriccoin.zcash.ui.screen.keystonebirthday.height
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class KeystoneBDHeightArgs(
+    val ur: String,
+    val accountIndex: Int
+)